reorganised the file structure
