+++ src/js/api/auth/login.js
@@ -1,6 +1,6 @@
 import { callApiWith } from "../apiCall.js";
 import { API_BASE, API_LOGIN } from "../../constants/index.js";
-import { save } from "../../storage/index.js";
+import { save } from "../../storage/_index.js";
 //add a call to fetch the api key??
 
 export async function login(email, password) {
