+++ src/js/api/auth/register.js
@@ -1,7 +1,7 @@
 import { callApiWith } from "../apiCall.js";
 import { API_BASE, API_REGISTER } from "../../constants/index.js";
-import { userFeedback } from "../../ui/components/errors/userFeedback.js";
 import { login } from "./index.js";
+import { userFeedback } from "../../ui/userFeedback/_index.js";
 
 let feedbackMessage;
 const errorContainer = document.getElementById("userFeedback");
