--- src/js/events/checkLoginStatus.js
@@ -1,22 +1,0 @@
-// import { load } from "../../storage/index.js";
-// import { logout } from "../../api/auth/index.js";
-
-// const token = load("token");
-// let pathname = "";
-
-// export function checkIfLoggedIn() {
-//   if (token) {
-//     pathname = window.location;
-//     return pathname;
-//   }
-//   if (!token) {
-//     logout();
-//     if (pathname.toLowerCase().includes("/semesterproject2_theaoland/")) {
-//       pathname = "/SemesterProject2_TheaOland/";
-//       return pathname;
-//     } else {
-//       pathname = "/";
-//       return pathname;
-//     }
-//   }
-// }
