+++ src/js/handlers/events/_index.js
@@ -1,0 +1,8 @@
+export * from "./formValidation.js";
+export * from "./generateFormFields.js";
+export * from "./generateImgInputs.js";
+export * from "./navBack.js";
+export * from "./pagination.js";
+export * from "./populateEditForm.js";
+export * from "./renderListings.js";
+export * from "./search.js";
