+++ src/js/handlers/events/search.js
@@ -1,8 +1,7 @@
-import { startFeed } from "../routes/feed.js";
-import { userFeedback } from "../ui/components/errors/userFeedback.js";
-import { updateTotalPageDisplay, updatePaginationBtns } from "./pagination.js";
-import { renderListings } from "../templates/renderListings.js";
-import { get } from "../api/requests/get.js";
+import { get } from "../../api/requests/get.js";
+import { startFeed } from "../../routes/feed.js";
+import { renderListings, updateTotalPageDisplay, updatePaginationBtns } from "./_index.js";
+import { userFeedback } from "../../ui/userFeedback/_index.js";
 
 let query;
 let numberOfListings;
