+++ src/js/handlers/listners/_index.js
@@ -1,0 +1,8 @@
+export * from "./authentication.js";
+export * from "./bids.js";
+export * from "./edit.js";
+export * from "./logout.js";
+export * from "./pageTurn.js";
+export * from "./search.js";
+export * from "./updateProfile.js";
+export * from "./validation.js";
