+++ src/js/storage/_index.js
@@ -1,3 +1,3 @@
 export * from "./load.js";
-export * from "./save.js";
 export * from "./remove.js";
+export * from "./save.js";
