+++ src/js/templates/_index.js
@@ -1,0 +1,7 @@
+export * from "./bids.js";
+export * from "./btns.js";
+export * from "./listing.js";
+export * from "./listings.js";
+export * from "./media.js";
+export * from "./nav.js";
+export * from "./profile.js";
