+++ src/js/templates/bids.js
@@ -1,20 +1,19 @@
-import { countdownTimer } from "../events/countdownTimer.js";
-import { publishNewBid } from "../api/requests/post.js";
+import { countdownTimer } from "../ux/components/countdownTimer.js";
 
-export function listenForNewBid(listingID) {
-  document.forms.placeBid.addEventListener("submit", (event) => {
-    event.preventDefault();
+// export function listenForNewBid(listingID) {
+//   document.forms.placeBid.addEventListener("submit", (event) => {
+//     event.preventDefault();
 
-    const form = event.target;
-    const formData = new FormData(form);
-    const newBid = Object.fromEntries(formData.entries());
+//     const form = event.target;
+//     const formData = new FormData(form);
+//     const newBid = Object.fromEntries(formData.entries());
 
-    let parsedBid = { ...newBid };
-    parsedBid.amount = parseInt(newBid.amount);
+//     let parsedBid = { ...newBid };
+//     parsedBid.amount = parseInt(newBid.amount);
 
-    publishNewBid(listingID, parsedBid);
-  });
-}
+//     publishNewBid(listingID, parsedBid);
+//   });
+// }
 
 export function addCurrentBid(bidsArray) {
   const bidArraylength = bidsArray.length;
