--- src/js/templates/index.js
@@ -1,2 +1,0 @@
-export * from "./listings.js";
-export * from "./nav.js";
