+++ src/js/ui/userFeedback/clearFeedback.js
